Move the flash and monitor wire type modules into probe-rs-rpc
