+++ probe-rs-rpc/Cargo.toml
@@ -12,7 +12,7 @@ rust-version.workspace = true
 
 [features]
 default = []
-clap = ["dep:clap"]
+clap = ["dep:clap", "dep:parse_int"]
 remote = ["dep:tokio-tungstenite", "dep:futures-util", "dep:axum"]
 
 [dependencies]
@@ -26,6 +26,7 @@ regex = "1.10.4"
 tokio = { version = "1.0", features = ["full"] }
 tokio-util = { version = "0.7", features = ["codec"] }
 clap = { version = "4", features = ["derive", "env"], optional = true }
+parse_int = { version = "0.9", optional = true }
 tokio-tungstenite = { version = "0.30", features = [
     "rustls-tls-webpki-roots",
 ], optional = true }
