+++ probe-rs-rpc/src/lib.rs
@@ -130,8 +130,13 @@ pub mod breakpoints;
 pub mod chip;
 pub mod core_ops;
 pub mod file;
+pub mod flash;
+pub mod format;
 pub mod memory;
+pub mod monitor;
 pub mod probe;
 pub mod reset;
 pub mod resume;
+pub mod rtt_client;
 pub mod rtt_config;
+pub mod semihosting_options;
